+++ application/classes/Ushahidi/Validator/Form/Create.php
@@ -8,28 +8,18 @@
  * @copyright  2014 Ushahidi
  * @license    https://www.gnu.org/licenses/agpl-3.0.html GNU Affero General Public License Version 3 (AGPL3)
  */
+use Ushahidi\Core\Entity;
+use Ushahidi\Core\Entity\FormRepository;
+use Ushahidi\Core\Tool\Validator;
+
 
 class Ushahidi_Validator_Form_Create extends Ushahidi_Validator_Form_Update
 {
 	protected function getRules()
 	{
 		return array_merge_recursive(parent::getRules(), [
-			'name' => [['not_empty']],
+			'name' => [['not_empty'],
 			[[$this, 'checkPostTypeLimit'], [':validation']],
-		]);
-	}
-
-	public function checkPostTypeLimit (Validation $validation)
-	{
-		$config = \Kohana::$config->load('features.limits');
-
-		if ($config['forms'] !== TRUE) {
-
-			$total_forms = $this->repo->getTotalCount();
-
-			if ($total_forms >= $config['forms']) {
-				$validation->error('name', 'postTypeLimitReached');
-			}
-		}
+		]]);
 	}
 }
